refactor(database): synchronize repositories and migrations
Synchronize database infrastructure with PostgreSQL migration:

Migrations:
- Restore 00002_create_base_tables.sql as compatibility NO-OP migration
- Preserve migration sequence history for existing installations
- Update all migrations for PostgreSQL syntax (BIGSERIAL, VARCHAR, NUMERIC, TIMESTAMP)
- Add composite indexes for performance optimization
- Add foreign key ON DELETE constraints for data integrity
- Migrate money fields from NUMERIC to INT64 for precision

Repositories:
- Update GORM repositories for PostgreSQL compatibility
- Refactor connection configuration for PostgreSQL
- Update RabbitMQ configuration
- Refactor category, company, dashboard, invitation, product, report, user repositories
- Update outbox repository tests
- Update stock movement repository tests

Ports:
- Update invitation repository interface
- Update user repository interface

Services:
- Refactor event dispatcher
- Refactor invitation service
- Refactor platform service
- Refactor purchase service

Handlers:
- Refactor company settings handler
- Refactor impersonation handler

Tests:
- Update concurrency tests
- Update tenant middleware tests

+++ backend/migrations/00006_add_processing_fields_stock_adjustments.sql
@@ -3,7 +3,7 @@
 
 -- Adicionar campos para rastreabilidade operacional
 ALTER TABLE stock_adjustments_pending 
-ADD COLUMN processed_by INTEGER;
+ADD COLUMN processed_by BIGINT;
 
 ALTER TABLE stock_adjustments_pending 
 ADD COLUMN processing_notes TEXT;
